formated separate files are created for each due to code length
