+++ app/controllers/interviewee/profile/index.js
@@ -1,0 +1,18 @@
+const express = require('express');
+const secondaryRoutes = require('./secondary');
+const seniorSecondaryRoutes = require('./seniorSecondary');
+const degreeRoutes = require('./degree');
+
+const router = express.Router();
+
+router.use('/secondary', secondaryRoutes);
+router.use('/seniorSecondary', seniorSecondaryRoutes);
+router.use('/degree', degreeRoutes);
+// router.use('/profession', professionRoutes);
+// router.use('/test', testRoutes);
+// router.use('/certification', certificationRoutes);
+// router.use('/course', courseRoutes);
+// router.use('/additional', additionalRoutes);
+// router.use('/skill', skillRoutes);
+// router.use('/project', projectRoutes);
+module.exports = router;
